+++ changelog.md
@@ -2,7 +2,9 @@ latest
  - Nudge Crossings selected action, nudges all zero crossings off from zero.
  - Pad with Zero selected action, adds a zero sample to the beginning and end of each file.
  - Shift selected action, moves the sample start position over 50%.
- - Support for reading slice data from the Sonicware LoFi 12XT custom header data.
+ - Fixed bug where resampling the list when changing the working samplerate caused the loop point to bet set on .ot slice metadata.
+ - Support for reading slice data from the Sonicware LoFi-12 XT custom header data.
+ - Support for writing slice data into wav files for use on the Sonicware LoFi-12 XT (in 12/24 kHz target sample rate).
 
 2024.05.28 1.4.15
  - Amplitude threshold condense selected action, enter an upper and lower amplitude value (between 0 and 1), the sample will be modified to only include amplitudes within the given range, gaps between the audio is removed condensing the sample size.